+++ packages/nx-terraform-plugin/src/fs.ts
@@ -2,16 +2,21 @@ import { RawProjectGraphDependency } from "@nx/devkit";
 import fs from "node:fs/promises";
 
 import { getStaticDependencies } from "./hcl.ts";
+import { getPackageLogger } from "./logger.ts";
 import { ProjectFile } from "./project-file.ts";
 
 export const getStaticDependenciesFromFile = async (
   file: ProjectFile,
 ): Promise<RawProjectGraphDependency[]> => {
+  const logger = getPackageLogger(["fs"]);
   try {
     const fileContent = await fs.readFile(file.fileName, "utf-8");
     return getStaticDependencies(file, fileContent);
   } catch (error) {
-    console.error(`Error reading file ${file.fileName}:`, error);
+    logger.error("Error reading file {fileName}", {
+      error,
+      fileName: file.fileName,
+    });
     return [];
   }
 };